feature-issuance module refactor, updated baseline profiles
