+++ issuance-feature/src/main/java/eu/europa/ec/issuancefeature/ui/success/DocumentIssuanceSuccessViewModel.kt
@@ -14,7 +14,7 @@
  * governing permissions and limitations under the Licence.
  */
 
-package eu.europa.ec.issuancefeature.ui.document.success
+package eu.europa.ec.issuancefeature.ui.success
 
 import androidx.lifecycle.viewModelScope
 import eu.europa.ec.commonfeature.config.IssuanceSuccessUiConfig
